backend/requirements.txt
# Backend Dependencies

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# AI/ML
openai==1.10.0
anthropic==0.8.1
langchain==0.1.4
tiktoken==0.5.2

# Data processing
numpy==1.26.3
pandas==2.2.0

# HTTP client
aiohttp==3.9.1
httpx==0.26.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Market data
alpha-vantage==2.3.1
yfinance==0.2.35

# CORS
python-cors==1.0.0
